refactor: tratar disciplinas

+++ backend/src/main/java/com/denisson/backend/domain/useCases/disciplinas/DeleteDisciplinaUseCase.java
@@ -1,5 +1,7 @@
 package com.denisson.backend.domain.useCases.disciplinas;
 
+import com.denisson.backend.domain.exceptions.NotFoundException;
+import com.denisson.backend.domain.models.Disciplina;
 import com.denisson.backend.domain.repositories.IDisciplinaRepository;
 
 public class DeleteDisciplinaUseCase {
@@ -11,6 +13,11 @@ public class DeleteDisciplinaUseCase {
     }
 
     public void execute(Long id) {
+        Disciplina existsDisciplina = repository.findById(id);
+        if (existsDisciplina == null) {
+            throw new NotFoundException("Disciplina não encontrada com id " + id);
+        }
+
         repository.delete(id);
     }
 }
